feat: refactor queue manager into ui layer
